# bpo-47005: Improve performance of bytearray_repeat and bytearray_irepeat