The `bytearray_repeat` and `bytearray_irepeat` are inefficient for small arrays and a high number of repeats.
This can be improved by using the same approach is in the corresponding `bytes_repeat` method.

Microbenchmarks:

```
python -m pyperf timeit "b=bytearray([1,2,])*100" 
python -m pyperf timeit "b=bytearray([1,2,])*1000"
```` 
Results:
```
Mean +- std dev: [base100] 479 ns +- 29 ns -> [patch100] 274 ns +- 18 ns: 1.75x faster
Mean +- std dev: [base1000] 2.58 us +- 0.18 us -> [patch1000] 399 ns +- 26 ns: 6.46x faster
```

https://bugs.python.org/issue47005

<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
bpo-NNNN: Summary of the changes made
```

Where: bpo-NNNN refers to the issue number in the https://bugs.python.org.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->


<!-- issue-number: [bpo-47005](https://bugs.python.org/issue47005) -->
https://bugs.python.org/issue47005
<!-- /issue-number -->